# microservice/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  comment.Comment:
    properties:
      be_replied_content:
        type: string
      be_replied_user_id:
        type: integer
      content:
        type: string
      create_time:
        type: string
      creator_avatar:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      father_id:
        type: integer
      id:
        type: integer
      img_url:
        type: string
      is_liked:
        type: boolean
      like_num:
        type: integer
      type_name:
        description: first-level -> 一级评论; second-level -> 其它级
        type: string
    type: object
...
  post.ListMainPostResponse:
    properties:
      posts:
        items:
          $ref: '#/definitions/post.Post'
        type: array
    type: object
  post.Post:
    properties:
      category:
        type: string
      collection_num:
        type: integer
      comment_num:
        type: integer
      comments:
        items:
          $ref: '#/definitions/comment.Comment'
        type: array
      content_type:
        description: md or rtf
        type: string
      creator_avatar:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      id:
        type: integer
      is_collection:
        type: boolean
      is_liked:
        type: boolean
      like_num:
        type: integer
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      time:
        type: string
      title:
        type: string
    type: object
  post.PostPartInfo:
    properties:
      category:
        type: string
      collection_num:
        type: integer
      comment_num:
        type: integer
      creator_avatar:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      id:
        type: integer
      like_num:
        type: integer
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      time:
        type: string
      title:
        type: string
    type: object
  post.PostPartInfoResponse:
    properties:
      posts:
        items:
          $ref: '#/definitions/post.PostPartInfo'
        type: array
    type: object
...
  report.CreateRequest:
    properties:
      category:
        description: 可选参数
        type: string
      cause:
        type: string
      id:
        description: post的id或者是comment的id
        type: integer
      type_name:
        description: post or comment
        type: string
    required:
    - cause
    - id
    - type_name
    type: object
  report.HandleRequest:
    properties:
      id:
        type: integer
      result:
        description: invalid or valid
        type: string
    required:
    - id
    - result
    type: object
  report.ListResponse:
    properties:
      reports:
        items:
          $ref: '#/definitions/report.Report'
        type: array
    type: object
  report.Report:
    properties:
      be_reported_content:
        type: string
      be_reported_user_id:
        type: integer
      be_reported_user_name:
        type: string
      category:
        type: string
      cause:
        type: string
      create_time:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      target_id:
        type: integer
      type_name:
        type: string
      user_avatar:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
  user.CreateMessageRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  user.CreatePrivateMessageRequest:
    properties:
      comment_id:
        type: integer
      content:
        type: string
      post_id:
        type: integer
      receive_userid:
        type: integer
      type:
        description: comment/like/collection/reply_comment
        type: string
    required:
    - post_id
    - receive_userid
    - type
    type: object
...
host: forum.muxixyz.com
info:
  contact: {}
  description: The gateway of forum
  title: github.com/Muxi-X/forum-be/microservice/gateway
  version: "1.0"
paths:
...
  /post/list/{domain}:
    get:
      consumes:
      - application/json
      description: 根据category or tag 获取主帖list
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: last_id
        in: query
        name: last_id
        type: integer
      - description: category
        in: query
        name: category
        type: string
      - description: filter
        in: query
        name: filter
        type: string
      - description: search_content
        in: query
        name: search_content
        type: string
      - description: tag
        in: query
        name: tag
        type: string
      - description: normal -> 团队外; muxi -> 团队内
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.ListMainPostResponse'
      summary: list 主帖 api
      tags:
      - post
...
  /post/published/{user_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: last_id
        in: query
        name: last_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostPartInfoResponse'
      summary: list 用户发布的帖子 api
      tags:
      - post
...
  /report:
    post:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: create_report_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/report.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 举报帖子 api
      tags:
      - report
    put:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: handle_report_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/report.HandleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 处理举报 api
      tags:
      - report
  /report/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: last_id
        in: query
        name: last_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.ListResponse'
      summary: list举报 api
      tags:
      - report
...
  /user/message:
    post:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: create_message_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/user.CreateMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 创建 公共 message api
      tags:
      - user
...
  /user/private_message:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 删除 个人 message api
      tags:
      - user
    post:
      consumes:
      - application/json
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: create_private_message_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/user.CreatePrivateMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 创建 个人 message api
      tags:
      - user
...
swagger: "2.0"
...